refactor: 💡 migrate to src directory
