+++ src/components/PlayerStyle.tsx
