+++ src/components/__tests__/FizzBuzz.test.tsx
@@ -1,3 +1,6 @@
+/**
+ * @jest-environment jsdom
+ */
 import React from 'react'
 import { cleanup, fireEvent, render } from '@testing-library/react'
 import FizzBuzz from '../FizzBuzz'
