+++ src/components/header/atoms/Left.stories.tsx
