+++ src/components/id/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import ReactMarkdown from 'react-markdown'
 import styled from 'styled-components'
-import { ContainerProps as PostProps } from '../../components/Post'
+import { ContainerProps as PostProps } from '../Post'
 import Layout from '../Layout'
 import DeleteButton from './atoms/DeleteButton'
 import PublishButton from './atoms/PublishButton'
