+++ src/lib/analytics.ts
