+++ src/lib/types/google-analytics/Event.ts
